+++ djangocms_snippet/cms_plugins.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from django import template
-from django.template.context import Context
+from django.conf import settings
 from django.utils.html import escape
 from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_lazy as _
@@ -11,37 +11,41 @@ from cms.plugin_pool import plugin_pool
 from .models import SnippetPtr
 
 
+CACHE_ENABLED = getattr(settings, 'DJANGOCMS_SNIPPET_CACHE', False)
+
+
 class SnippetPlugin(CMSPluginBase):
     model = SnippetPtr
     name = _('Snippet')
     render_template = 'djangocms_snippet/snippet.html'
     text_enabled = True
     text_editor_preview = False
-    cache = getattr(settings, 'DJANGOCMS_SNIPPET_CACHE', True)
+    cache = CACHE_ENABLED
 
     def render(self, context, instance, placeholder):
+        try:
+            if instance.snippet.template:
+                context = context.flatten()
+                t = template.loader.get_template(instance.snippet.template)
+                content = t.render(context)
+            else:
+                # only html provided
+                t = template.Template(instance.snippet.html)
+                content = t.render(context)
+        except template.TemplateDoesNotExist:
+            content = _('Template %(template)s does not exist.') % {
+                'template': instance.snippet.template
+            }
+        except Exception as e:
+            content = escape(str(e))
+
         context.update({
             'placeholder': placeholder,
             'object': instance,
+            'html': mark_safe(instance.snippet.html),
+            'content': content,
         })
-        try:
-            if instance.snippet.template:
-                t = template.loader.get_template(instance.snippet.template)
-                context.update({
-                    'html': mark_safe(instance.snippet.html)
-                })
-                content = t.render(Context(context))
-            else:
-                t = template.Template(instance.snippet.html)
-                content = t.render(Context(context))
-        except template.TemplateDoesNotExist:
-            content = _('Template %(template)s does not exist.') % {
-                'template': instance.snippet.template}
-        except Exception as e:
-            content = escape(str(e))
-        context.update({
-            'content': mark_safe(content),
-        })
+
         return context