+++ CHANGELOG.rst
@@ -3,6 +3,13 @@ Changelog
 =========
 
 
+2.3.0 (unreleased)
+==================
+
+* Fixed an issue where render requires a dict instead of a context
+* Added ``DJANGOCMS_SNIPPET_CACHE`` cache settings for snippets
+
+
 2.2.0 (2019-05-06)
 ==================